Move components folder outside
