+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import { AlertProvider } from "./components/contexts/alert-provider";
+import { AlertProvider } from "../components/contexts/alert-provider";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
